streamlit_deployment/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
cohere>=4.0.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0 